refactor: code refactor & add docs
