+++ src/components/views/WordpressMigration/DefaultRootView.client.tsx
@@ -4,7 +4,6 @@ import { Button, Gutter } from '@payloadcms/ui'
 import React, { useCallback, useState } from 'react'
 
 export const DefaultRootViewClient = ({ user }: { user: any }) => {
-  console.log('user', user)
   const [file, setFile] = useState<File | null>(null)
   const [isLoading, setIsLoading] = useState(false)
   const [error, setError] = useState<string | null>(null)
